+++ src/index.js
@@ -4,6 +4,7 @@ import EsButtonGroup from "../packages/button-group/index"
 import EsCrumb from "../packages/crumb/index"
 import EsMenu from "../packages/menu/index"
 import EsQuery from "../packages/query/index"
+import EsTable from "../packages/table/index"
 import EsTabs from "../packages/tabs/index"
 import 'element-ui/lib/theme-chalk/index.css'
 import '../lib/css/public.css'
@@ -15,6 +16,7 @@ const components = [
   EsCrumb,
   EsMenu,
   EsQuery,
+  EsTable,
   EsTabs
 ]